move Filter Value class to Filter::Base, to extend it in each filter type accordingly
